requirements.txt
anthropic>=0.42.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
sqlalchemy>=2.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
